change mysql-querier to singleton pattern
this allows mysql-querier to be ended at the top-level
since one instance is shared

before this change, new instances of mysql-querier were created
at each call to require().  thus, mysql-querier needed to be
deleted at the end of each use, which is not the desired behaviour.

no change to depending classes is necessary.